+++ pkg/logs/tailers/windowsevent/tailer.go
