+++ pkg/logs/tailers/windowsevent/tailer_test.go
