+++ pkg/logs/tailers/windowsevent/transform.go
