+++ pkg/logs/tailers/windowsevent/transform_test.go
