+++ tasks/agent.py
@@ -484,9 +484,9 @@ def _windows_integration_tests(ctx, race=False, go_mod="mod", arch="x64"):
         },
         {
             # Run eventlog tailer tests with the Windows API, which depend on the EventLog service
-            'prefix': './pkg/logs/internal/tailers/windowsevent-new/...',
+            'prefix': './pkg/logs/tailers/windowsevent/...',
             'extra_args': '-evtapi Windows',
-        }
+        },
     ]
 
     for test in tests:
